完善小码流视频数据配置参数&设备内存以及CPU占用率保留2位小数&帧率通知采用整数上报新增下行整体丢包率统计

+++ VCSSDK/Depend/VCSSDK.framework/PrivateHeaders/VCSStreamRemote.h
@@ -20,7 +20,7 @@ NS_ASSUME_NONNULL_BEGIN
 /// 视频流高
 @property (assign, nonatomic, readonly) int height;
 /// 视频帧率
-@property (assign, nonatomic, readonly) float fps;
+@property (assign, nonatomic, readonly) int fps;
 /// 最后一次帧处理时间戳(毫秒)
 @property (assign, nonatomic, readonly) NSTimeInterval interval;
